Comparison PrimFn's now return long's as opposed to int's

+++ src/main/java/org/idris/rts/PrimFn.java
@@ -680,25 +680,25 @@ public class PrimFn {
         return res;
     }
     
-    public static int LEq(byte x, byte y) {
+    public static long LEq(byte x, byte y) {
         return (x == y ? 1 : 0);
     }
-    public static int LEq(char x, char y) {
+    public static long LEq(char x, char y) {
         return (x == y ? 1 : 0);
     }
-    public static int LEq(short x, short y) {
+    public static long LEq(short x, short y) {
         return (x == y ? 1 : 0);
     }
-    public static int LEq(int x, int y) {
+    public static long LEq(int x, int y) {
         return (x == y ? 1 : 0);
     }
-    public static int LEq(long x, long y) {
+    public static long LEq(long x, long y) {
         return (x == y ? 1 : 0);
     }
-    public static int LEq(double x, double y) {
+    public static long LEq(double x, double y) {
         return (x == y ? 1 : 0);
     }
-    public static int LEq(BigInteger x, BigInteger y) {
+    public static long LEq(BigInteger x, BigInteger y) {
         return (x.equals(y) ? 1 : 0);
     }
     public static byte[] LEq(byte[] x, byte[] y) {
@@ -730,25 +730,25 @@ public class PrimFn {
         return res;
     }
     
-    public static int LSLt(byte x, byte y) {
+    public static long LSLt(byte x, byte y) {
         return (x < y ? 1 : 0);
     }
-    public static int LSLt(char x, char y) {
+    public static long LSLt(char x, char y) {
         return (x < y ? 1 : 0);
     } 
-    public static int LSLt(short x, short y) {
+    public static long LSLt(short x, short y) {
         return (x < y ? 1 : 0);
     }
-    public static int LSLt(int x, int y) {
+    public static long LSLt(int x, int y) {
         return (x < y ? 1 : 0);
     }
-    public static int LSLt(long x, long y) {
+    public static long LSLt(long x, long y) {
         return (x < y ? 1 : 0);
     }
-    public static double LSLt(double x, double y) {
+    public static long LSLt(double x, double y) {
         return (x < y ? 1 : 0);
     }
-    public static int LSLt(BigInteger x, BigInteger y) {
+    public static long LSLt(BigInteger x, BigInteger y) {
         return (x.compareTo(y) < 0 ? 1 : 0);
     }
     public static byte[] LSLt(byte[] x, byte[] y) {
@@ -779,19 +779,19 @@ public class PrimFn {
         }
         return res;
     }
-    public static int LLt(byte x, byte y) {
+    public static long LLt(byte x, byte y) {
         return ((x < y) ^ (x < 0) ^ (y < 0)  ? 1 : 0);
     }
-    public static int LLt(char x, char y) {
+    public static long LLt(char x, char y) {
         return ((x < y) ^ (x < 0) ^ (y < 0)  ? 1 : 0);
     } 
-    public static int LLt(short x, short y) {
+    public static long LLt(short x, short y) {
         return ((x < y) ^ (x < 0) ^ (y < 0)  ? 1 : 0);
     }
-    public static int LLt(int x, int y) {
+    public static long LLt(int x, int y) {
         return ((x < y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LLt(long x, long y) {
+    public static long LLt(long x, long y) {
         return ((x < y) ^ (x < 0l) ^ (y < 0l) ? 1 : 0);
     }
     public static byte[] LLt(byte[] x, byte[] y) {
@@ -823,25 +823,25 @@ public class PrimFn {
         return res;
     }
     
-    public static int LSLe(byte x, byte y) {
+    public static long LSLe(byte x, byte y) {
         return (x <= y ? 1 : 0);
     }
-    public static int LSLe(char x, char y) {
+    public static long LSLe(char x, char y) {
         return (x <= y ? 1 : 0);
     }
-    public static int LSLe(short x, short y) {
+    public static long LSLe(short x, short y) {
         return (x <= y ? 1 : 0);
     }
-    public static int LSLe(int x, int y) {
+    public static long LSLe(int x, int y) {
         return (x <= y ? 1 : 0);
     }
-    public static int LSLe(long x, long y) {
+    public static long LSLe(long x, long y) {
         return (x <= y ? 1 : 0);
     }
-    public static double LSLe(double x, double y) {
+    public static long LSLe(double x, double y) {
         return (x <= y ? 1 : 0);
     }
-    public static int LSLe(BigInteger x, BigInteger y) {
+    public static long LSLe(BigInteger x, BigInteger y) {
         return (x.compareTo(y) <= 0 ? 1 : 0);
     }
     public static byte[] LSLe(byte[] x, byte[] y) {
@@ -873,19 +873,19 @@ public class PrimFn {
         return res;
     }
     
-    public static int LLe(byte x, byte y) {
+    public static long LLe(byte x, byte y) {
         return ((x <= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LLe(char x, char y) {
+    public static long LLe(char x, char y) {
         return ((x <= y) ^ (x < 0l) ^ (y < 0l) ? 1 : 0);
     }
-    public static int LLe(short x, short y) {
+    public static long LLe(short x, short y) {
         return ((x <= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LLe(int x, int y) {
+    public static long LLe(int x, int y) {
         return ((x <= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LLe(long x, long y) {
+    public static long LLe(long x, long y) {
         return ((x <= y) ^ (x < 0l) ^ (y < 0l) ? 1 : 0);
     }
     public static byte[] LLe(byte[] x, byte[] y) {
@@ -917,25 +917,25 @@ public class PrimFn {
         return res;
     }
     
-    public static int LSGt(byte x, byte y) {
+    public static long LSGt(byte x, byte y) {
         return (x > y ? 1 : 0);
     }    
-    public static int LSGt(char x, char y) {
+    public static long LSGt(char x, char y) {
         return (x > y ? 1 : 0);
     }  
-    public static int LSGt(short x, short y) {
+    public static long LSGt(short x, short y) {
         return (x > y ? 1 : 0);
     }
-    public static int LSGt(int x, int y) {
+    public static long LSGt(int x, int y) {
         return (x > y ? 1 : 0);
     }
-    public static int LSGt(long x, long y) {
+    public static long LSGt(long x, long y) {
         return (x > y ? 1 : 0);
     }
-    public static double LSGt(double x, double y) {
+    public static long LSGt(double x, double y) {
         return (x > y ? 1 : 0);
     }
-    public static int LSGt(BigInteger x, BigInteger y) {
+    public static long LSGt(BigInteger x, BigInteger y) {
         return (x.compareTo(y) > 0 ? 1 : 0);
     }
     public static byte[] LSGt(byte[] x, byte[] y) {
@@ -966,19 +966,19 @@ public class PrimFn {
         }
         return res;
     }
-    public static int LGt(byte x, byte y) {
+    public static long LGt(byte x, byte y) {
         return ((x > y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }    
-    public static int LGt(char x, char y) {
+    public static long LGt(char x, char y) {
         return ((x > y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }  
-    public static int LGt(short x, short y) {
+    public static long LGt(short x, short y) {
         return ((x > y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LGt(int x, int y) {
+    public static long LGt(int x, int y) {
         return ((x > y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LGt(long x, long y) {
+    public static long LGt(long x, long y) {
         return ((x > y) ^ (x < 0l) ^ (y < 0l) ? 1 : 0);
     }
     public static byte[] LGt(byte[] x, byte[] y) {
@@ -1010,25 +1010,25 @@ public class PrimFn {
         return res;
     }
     
-    public static int LSGe(byte x, byte y) {
+    public static long LSGe(byte x, byte y) {
         return (x >= y ? 1 : 0);
     }
-    public static int LSGe(char x, char y) {
+    public static long LSGe(char x, char y) {
         return (x >= y ? 1 : 0);
     }
-    public static int LSGe(short x, short y) {
+    public static long LSGe(short x, short y) {
         return (x >= y ? 1 : 0);
     }
-    public static int LSGe(int x, int y) {
+    public static long LSGe(int x, int y) {
         return (x >= y ? 1 : 0);
     }
-    public static int LSGe(long x, long y) {
+    public static long LSGe(long x, long y) {
         return (x >= y ? 1 : 0);
     }
-    public static double LSGe(double x, double y) {
+    public static long LSGe(double x, double y) {
         return (x >= y ? 1 : 0);
     }
-    public static int LSGe(BigInteger x, BigInteger y) {
+    public static long LSGe(BigInteger x, BigInteger y) {
         return (x.compareTo(y) >= 0 ? 1 : 0);
     }
     public static byte[] LSGe(byte[] x, byte[] y) {
@@ -1059,19 +1059,19 @@ public class PrimFn {
         }
         return res;
     }   
-    public static int LGe(byte x, byte y) {
+    public static long LGe(byte x, byte y) {
         return ((x >= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LGe(char x, char y) {
+    public static long LGe(char x, char y) {
         return ((x >= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LGe(short x, short y) {
+    public static long LGe(short x, short y) {
         return ((x >= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LGe(int x, int y) {
+    public static long LGe(int x, int y) {
         return ((x >= y) ^ (x < 0) ^ (y < 0) ? 1 : 0);
     }
-    public static int LGe(long x, long y) {
+    public static long LGe(long x, long y) {
         return ((x >= y) ^ (x < 0l) ^ (y < 0l) ? 1 : 0);
     }
     public static byte[] LGe(byte[] x, byte[] y) {
